refactor(runtimed): split peer writer module
