+++ crates/runtimed/src/notebook_sync_server/peer.rs
@@ -1,3 +1,7 @@
+use super::peer_writer::{
+    queue_request_error, queue_session_status, spawn_peer_request_worker, spawn_peer_writer,
+    PeerWriter, RequestEnqueueError,
+};
 use super::*;
 use runtime_doc::RuntimeLifecycle;
 
@@ -1108,205 +1112,6 @@ where
     )
     .await?;
     Ok(())
-}
-
-const PEER_OUTBOUND_QUEUE_CAPACITY: usize = 1024;
-const PEER_REQUEST_QUEUE_CAPACITY: usize = 64;
-
-struct OutboundFrame {
-    frame_type: NotebookFrameType,
-    payload: Vec<u8>,
-}
-
-#[derive(Clone)]
-struct PeerWriter {
-    tx: mpsc::Sender<OutboundFrame>,
-}
-
-struct PeerWriterTask {
-    handle: tokio::task::JoinHandle<anyhow::Result<()>>,
-}
-
-struct PeerRequestWorker {
-    tx: mpsc::Sender<notebook_protocol::protocol::NotebookRequestEnvelope>,
-    handle: tokio::task::JoinHandle<anyhow::Result<()>>,
-}
-
-#[derive(Debug)]
-enum RequestEnqueueError {
-    Full(notebook_protocol::protocol::NotebookRequestEnvelope),
-    Closed(notebook_protocol::protocol::NotebookRequestEnvelope),
-}
-
-impl Drop for PeerWriterTask {
-    fn drop(&mut self) {
-        self.handle.abort();
-    }
-}
-
-impl Drop for PeerRequestWorker {
-    fn drop(&mut self) {
-        self.handle.abort();
-    }
-}
-
-impl PeerWriter {
-    fn send_frame(&self, frame_type: NotebookFrameType, payload: Vec<u8>) -> anyhow::Result<()> {
-        self.tx
-            .try_send(OutboundFrame {
-                frame_type,
-                payload,
-            })
-            .map_err(|e| match e {
-                mpsc::error::TrySendError::Full(frame) => anyhow::anyhow!(
-                    "peer outbound queue full while sending {:?} frame",
-                    frame.frame_type
-                ),
-                mpsc::error::TrySendError::Closed(frame) => anyhow::anyhow!(
-                    "peer writer stopped before sending {:?} frame",
-                    frame.frame_type
-                ),
-            })
-    }
-
-    fn send_json<T>(&self, frame_type: NotebookFrameType, value: &T) -> anyhow::Result<()>
-    where
-        T: serde::Serialize,
-    {
-        let payload = serde_json::to_vec(value)?;
-        self.send_frame(frame_type, payload)
-    }
-
-    /// Number of free slots in the outbound channel.
-    ///
-    /// `PEER_OUTBOUND_QUEUE_CAPACITY - capacity()` gives the number of
-    /// in-flight frames waiting to be flushed to the socket — useful as a
-    /// backpressure signal in telemetry.
-    fn capacity(&self) -> usize {
-        self.tx.capacity()
-    }
-}
-
-impl PeerRequestWorker {
-    fn enqueue(
-        &self,
-        envelope: notebook_protocol::protocol::NotebookRequestEnvelope,
-    ) -> Result<(), RequestEnqueueError> {
-        self.tx.try_send(envelope).map_err(|e| match e {
-            mpsc::error::TrySendError::Full(envelope) => RequestEnqueueError::Full(envelope),
-            mpsc::error::TrySendError::Closed(envelope) => RequestEnqueueError::Closed(envelope),
-        })
-    }
-}
-
-fn spawn_peer_writer<W>(
-    mut writer: W,
-    notebook_id: String,
-    peer_id: String,
-) -> (PeerWriter, PeerWriterTask)
-where
-    W: AsyncWrite + Unpin + Send + 'static,
-{
-    let (tx, mut rx) = mpsc::channel::<OutboundFrame>(PEER_OUTBOUND_QUEUE_CAPACITY);
-    let handle = tokio::spawn(async move {
-        while let Some(frame) = rx.recv().await {
-            connection::send_typed_frame(&mut writer, frame.frame_type, &frame.payload)
-                .await
-                .map_err(|e| {
-                    anyhow::anyhow!(
-                        "failed to write {:?} frame to peer {} for {}: {}",
-                        frame.frame_type,
-                        peer_id,
-                        notebook_id,
-                        e
-                    )
-                })?;
-        }
-        Ok(())
-    });
-    (PeerWriter { tx }, PeerWriterTask { handle })
-}
-
-fn spawn_peer_request_worker(
-    room: Arc<NotebookRoom>,
-    daemon: std::sync::Arc<crate::daemon::Daemon>,
-    writer: PeerWriter,
-    notebook_id: String,
-    peer_id: String,
-) -> PeerRequestWorker {
-    let (tx, mut rx) = mpsc::channel::<notebook_protocol::protocol::NotebookRequestEnvelope>(
-        PEER_REQUEST_QUEUE_CAPACITY,
-    );
-    let handle = tokio::spawn(async move {
-        while let Some(envelope) = rx.recv().await {
-            let label = metadata::request_label(&envelope.request);
-            let req_id = envelope.id.as_deref().unwrap_or("-");
-            let writer_queue_depth = PEER_OUTBOUND_QUEUE_CAPACITY - writer.capacity();
-            debug!(
-                "[notebook-sync] Request {} id={} peer={} notebook={} writer_queue={}",
-                label, req_id, peer_id, notebook_id, writer_queue_depth,
-            );
-
-            let start = std::time::Instant::now();
-            let response = handle_notebook_request(&room, envelope.request, daemon.clone()).await;
-            let elapsed = start.elapsed();
-            debug!(
-                "[notebook-sync] Request {} id={} completed in {:?}",
-                label, req_id, elapsed,
-            );
-
-            let reply = notebook_protocol::protocol::NotebookResponseEnvelope {
-                id: envelope.id,
-                response,
-            };
-            writer
-                .send_json(NotebookFrameType::Response, &reply)
-                .map_err(|e| {
-                    anyhow::anyhow!(
-                        "failed to queue response to peer {} for {}: {}",
-                        peer_id,
-                        notebook_id,
-                        e
-                    )
-                })?;
-        }
-        Ok(())
-    });
-    PeerRequestWorker { tx, handle }
-}
-
-fn queue_request_error(
-    writer: &PeerWriter,
-    id: Option<String>,
-    error: impl Into<String>,
-) -> anyhow::Result<()> {
-    writer.send_json(
-        NotebookFrameType::Response,
-        &notebook_protocol::protocol::NotebookResponseEnvelope {
-            id,
-            response: notebook_protocol::protocol::NotebookResponse::Error {
-                error: error.into(),
-            },
-        },
-    )
-}
-
-fn queue_session_status(
-    writer: &PeerWriter,
-    notebook_doc: notebook_protocol::protocol::NotebookDocPhaseWire,
-    runtime_state: notebook_protocol::protocol::RuntimeStatePhaseWire,
-    initial_load: notebook_protocol::protocol::InitialLoadPhaseWire,
-) -> anyhow::Result<()> {
-    writer.send_json(
-        NotebookFrameType::SessionControl,
-        &notebook_protocol::protocol::SessionControlMessage::SyncStatus(
-            notebook_protocol::protocol::SessionSyncStatusWire {
-                notebook_doc,
-                runtime_state,
-                initial_load,
-            },
-        ),
-    )
 }
 
 /// State carried from the initial notebook-doc sync into the steady-state loop.
@@ -2466,82 +2271,4 @@ pub(crate) fn notebook_execution_context_id(
     notebook_path
         .map(str::to_string)
         .unwrap_or_else(|| room.id.to_string())
-}
-
-#[cfg(test)]
-mod peer_writer_tests {
-    use super::*;
-
-    #[tokio::test]
-    async fn peer_writer_enqueue_does_not_wait_for_socket_drain() {
-        let (server, client) = tokio::io::duplex(16);
-        let (writer, mut task) =
-            spawn_peer_writer(server, "notebook".to_string(), "peer".to_string());
-
-        writer
-            .send_frame(NotebookFrameType::AutomergeSync, vec![1; 4096])
-            .unwrap();
-        tokio::time::sleep(std::time::Duration::from_millis(10)).await;
-        assert!(
-            !task.handle.is_finished(),
-            "writer task should be backpressured by the socket"
-        );
-
-        for index in 0..32 {
-            writer
-                .send_frame(NotebookFrameType::Presence, vec![index])
-                .unwrap();
-        }
-
-        drop(writer);
-        drop(client);
-
-        let result = tokio::time::timeout(std::time::Duration::from_secs(1), &mut task.handle)
-            .await
-            .expect("writer task should observe the closed socket")
-            .expect("writer task should not panic");
-        assert!(
-            result.is_err(),
-            "closed socket should surface as a writer task error"
-        );
-    }
-
-    #[tokio::test]
-    async fn peer_request_enqueue_reports_full_without_waiting_for_worker() {
-        let (tx, mut rx) = mpsc::channel::<notebook_protocol::protocol::NotebookRequestEnvelope>(1);
-        let (started_tx, started_rx) = oneshot::channel();
-        let handle = tokio::spawn(async move {
-            let _first = rx.recv().await;
-            let _ = started_tx.send(());
-            std::future::pending::<anyhow::Result<()>>().await
-        });
-        let worker = PeerRequestWorker { tx, handle };
-
-        worker
-            .enqueue(notebook_protocol::protocol::NotebookRequestEnvelope {
-                id: Some("first".to_string()),
-                request: NotebookRequest::GetDocBytes {},
-            })
-            .expect("first request should enqueue");
-        started_rx.await.expect("worker should start first request");
-        worker
-            .enqueue(notebook_protocol::protocol::NotebookRequestEnvelope {
-                id: Some("second".to_string()),
-                request: NotebookRequest::GetDocBytes {},
-            })
-            .expect("second request should fill the queue");
-
-        let start = std::time::Instant::now();
-        let err = worker
-            .enqueue(notebook_protocol::protocol::NotebookRequestEnvelope {
-                id: Some("third".to_string()),
-                request: NotebookRequest::GetDocBytes {},
-            })
-            .expect_err("full queue should reject immediately");
-        assert!(
-            start.elapsed() < std::time::Duration::from_millis(50),
-            "request enqueue should not wait for the busy worker"
-        );
-        assert!(matches!(err, RequestEnqueueError::Full(_)));
-    }
 }
